+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-08-09
+
 ### Added
 
 - `Board`, `DetectMode` and `Version`: what a program is running on.
@@ -670,7 +672,8 @@ and this project adheres to
 - Dual MIT / Apache-2.0 licensing
 - A draft of the cross-compilation setup in `docs/cross-compile.md`
 
-[Unreleased]: https://github.com/akiomik/bela-rs/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/akiomik/bela-rs/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/akiomik/bela-rs/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/akiomik/bela-rs/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/akiomik/bela-rs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/akiomik/bela-rs/compare/v0.0.1...v0.1.0